add pw_distribute_method2.cpp
