+++ source/module_pw/pw_basis.h
@@ -81,7 +81,8 @@ public:
     void distribute_r();
 
     //distribute plane waves to different processors
-    void distribution_method1();
+    void distribution_method1(); // x varies fast
+    void distribution_method2(); // y varies fast
 
 
    
@@ -140,6 +141,38 @@ private:
         ModuleBase::Vector3<double>**gdirect2D       // the i^th row contains the direct coordinates of planewaves that belong to the i^th core.
     );
     void get_ig2isz_is2ixy(
+        const int tot_nst,  // total number of sticks.
+        int* st_i,          // x or x + nx (if x < 0) of stick.
+        int* st_j,          // y or y + ny (if y < 0) of stick.
+        int* st_bottom,     // minimum z of stick, stored in 1d array with tot_nst elements.
+        int* st_length,     // the stick on (x, y) consists of st_length[x*ny+y] planewaves.
+        int* is2ip          // ip of core containing is^th stick, map is to ip.
+    );
+
+    void divide_sticks2(
+        const int tot_npw,  // total number of planewaves.
+        const int tot_nst,  // total number of sticks.
+        int* st_i,          // x or x + nx (if x < 0) of stick.
+        int* st_j,          // y or y + ny (if y < 0) of stick.
+        int* st_length,     // the stick on (x, y) consists of st_length[x*ny+y] planewaves.
+        int* npw_per,       // number of planewaves on each core.
+        int* nst_per,       // number of sticks on each core.
+        int* is2ip         // ip of core containing is^th stick, map is to ip.         
+    );
+    void get_istot2ixy2(
+        const int tot_nst,  // total number of sticks.
+        int* st_i,          // x or x + nx (if x < 0) of stick.
+        int* st_j,          // y or y + ny (if y < 0) of stick.
+        int* is2ip          // ip of core containing is^th stick, map is to ip.
+    );
+    void divide_pw2(
+        const int tot_npw,                          // total number of planewaves.
+        double* gg_global,                          // the modulus of all planewaves.
+        ModuleBase::Vector3<double>*gdirect_global, // direct coordinates of planewaves.
+        double** gg2D,                               // the i^th row contains the modulus of planewaves that belong to the i^th core.
+        ModuleBase::Vector3<double>**gdirect2D       // the i^th row contains the direct coordinates of planewaves that belong to the i^th core.
+    );
+    void get_ig2isz_is2ixy2(
         const int tot_nst,  // total number of sticks.
         int* st_i,          // x or x + nx (if x < 0) of stick.
         int* st_j,          // y or y + ny (if y < 0) of stick.
